Implemented getting profile from Discogs
fullupdate() now also pulls the user's profile from Discogs. A variant Profile was added to ParseType and the build_url() function to support this.
src/screens/profile.rs will implement a different profile struct, one with local app data (most listened to album, most expensive album, etc).

+++ src/app/mod.rs
@@ -27,7 +27,7 @@ pub struct Profile {
     pub collection: u64,
     pub wantlist: u64,
     pub rated: u64,
-    pub average_rating: u64,
+    pub average_rating: f64,
 }
 
 #[derive(Debug, Clone)]
